db-auto-scaling: moved config interface to lib folder
