+++ packages/db-auto-scaling/lib/IDbAutoScalingConfig.ts
