+++ packages/db/lib/IDbConfig.ts
