+++ packages/events/lib/IEventEmitterConfig.ts
